Update to include Jobs-Springboot3 in the project.
New steps and configurations are added for Jobs-Springboot3 and an existing file is deleted from Jobs-Springboot2. Jobs-Springboot3 is now included in the project modules in pom.xml and new files needed for its functionality are added. JobDescriptor, JobService, and multiple Dto class files are included. Also, existing JobStateDto file in Jobs-Springboot2 was updated. These changes combine to start the integration of Jobs-Springboot3 into the overall project structure.